requirements.txt
-e .
azureml-core
azureml-dataset-runtime
hydra-core
kornia
matplotlib
numpy
joblib
torch==1.9.0
pytorch-lightning
python-dotenv
torch-enhance
torchmetrics
torchvision
wandb
hydra-optuna-sweeper